requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.30,<1.0
pymongo>=4.8,<5.0
python-dotenv>=1.0,<2.0
jinja2>=3.1,<4.0
certifi>=2024.0.0,<2027.0.0
